requirements.txt
Django==2.2.10
djangorestframework==3.10.2

